serviceRegistryCommands: activate the left side of the window
