+++ packages/serverless-workflow-diagram-editor-envelope/src/envelope/ServerlessWorkflowDiagramEditorFactory.ts
@@ -34,8 +34,7 @@ export const onNodeClicked =
     if (!nodeName) {
       return;
     }
-    /* TODO: ServerlessWorkflowDiagramEditorFactory: replicate this to VSCode */
-    ctx.channelApi.notifications.kogitoSwfServiceCatalog_nodeClicked.send(nodeName);
+    ctx.channelApi.notifications.kogitoSwfServiceCatalog_moveCursorToNode.send(nodeName);
   };
 
 export class ServerlessWorkflowDiagramEditorFactory
