+++ packages/serverless-workflow-service-catalog/src/api/SwfServiceCatalogChannelApi.ts
@@ -28,5 +28,5 @@ export interface SwfServiceCatalogChannelApi {
     documentUri: string;
   }): void;
   kogitoSwfServiceCatalog_setupServiceRegistryUrl(): void;
-  kogitoSwfServiceCatalog_nodeClicked(nodeName: string): void;
+  kogitoSwfServiceCatalog_moveCursorToNode(nodeName: string): void;
 }
