+++ packages/serverless-workflow-text-editor/src/editor/textEditor/index.ts
@@ -16,4 +16,3 @@
 
 export * from "./SwfTextEditor";
 export * from "./SwfTextEditorController";
-export * from "./Utils";
